# src/main/java/com/ibm/api/BqPaymentMechanismApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.ibm.api;

import com.ibm.model.HTTPError;
import com.ibm.model.RetrievePaymentMechanismResponse;
import io.swagger.annotations.*;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import jakarta.validation.constraints.*;
@Validated
@Api(value = "BqPaymentMechanism", description = "the BqPaymentMechanism API")
public interface BqPaymentMechanismApi {
    
/**
 * ReBQ Retrieve details about the payment mechanism selection
 * @param paymentexecutionid.
 * @param paymentmechanismid.
 * @return A ResponseEntity containing the RetrievePaymentMechanismResponse object.
 */

  @ApiOperation(value = "ReBQ Retrieve details about the payment mechanism selection", nickname = "retrievePaymentMechanism", notes = "ReBQ Retrieve details about the payment mechanism selection", response = RetrievePaymentMechanismResponse.class , tags = { "BQ - PaymentMechanism" }  )
    @ApiResponses({ @ApiResponse(code = 200, message = "RetrievePaymentMechanismResponse", response = RetrievePaymentMechanismResponse.class),@ApiResponse(code = 400, message = "BadRequest", response = HTTPError.class),@ApiResponse(code = 401, message = "Unauthorized", response = HTTPError.class),@ApiResponse(code = 403, message = "Forbidden", response = HTTPError.class),@ApiResponse(code = 404, message = "NotFound", response = HTTPError.class),@ApiResponse(code = 429, message = "TooManyRequests", response = HTTPError.class),@ApiResponse(code = 500, message = "InternalServerError", response = HTTPError.class) })
     @RequestMapping(method = RequestMethod.GET,
            value = "/PaymentExecution/{paymentexecutionid}/PaymentMechanism/{paymentmechanismid}/Retrieve",
            produces = { "application/json" })
    ResponseEntity<RetrievePaymentMechanismResponse> retrievePaymentMechanism(@Parameter(name = "paymentexecutionid", description = "", required = true, in = ParameterIn.PATH) @PathVariable("paymentexecutionid") String paymentexecutionid,@Parameter(name = "paymentmechanismid", description = "", required = true, in = ParameterIn.PATH) @PathVariable("paymentmechanismid") String paymentmechanismid);
      
}